@@ -17,8 +17,7 @@
 		]
 	},
 	"prettier": {
-		"singleQuote": true,
-		"tab-width": 2
+		"singleQuote": true
 	},
 	"workspaces": [
 		"packages/*"
@@ -32,14 +31,13 @@
 		"lerna": "^2.5.1",
 		"mocha": "^4.1.0",
 		"rimraf": "^2.6.2",
-		"rxjs-marbles": "^2.3.0",
-		"ts-node": "^5.0.1",
-		"tslint": "5.9.1",
-		"typescript": "^2.8.1",
+		"ts-node": "^6.1.2",
+		"tslint": "5.10.0",
+		"typescript": "^2.9.2",
 		"sinon": "4.1.3",
 		"tsconfig-paths": "2.6.0",
-		"prettier": "1.12.0",
-		"lint-staged": "6.0.0",
+		"prettier": "1.13.5",
+		"lint-staged": "7.2.0",
 		"husky": "0.14.3",
 		"bignumber.js": "6.0.0",
 		"web3": "^0.20.6"
