@@ -1,8 +1,5 @@
-import { Observable } from 'rxjs/Observable';
+import { Observable, NEVER, of, merge } from 'rxjs';
 import { tap, finalize, mergeMapTo } from 'rxjs/operators';
-import { never } from 'rxjs/observable/never';
-import { of } from 'rxjs/observable/of';
-import { merge } from 'rxjs/observable/merge';
 
 import { getQueryResultsFromModelId } from '../store';
 import * as fromEvents from '../modules/events';
@@ -14,7 +11,7 @@ export const query = ({ genId, options, store }: Context) => (
 ): Observable<any> => {
   const id = genId();
   return merge(
-    never(),
+    NEVER,
     of(queryModel).pipe(
       getInterceptor('preQuery', options),
       tap(() =>
