@@ -1,4 +1,4 @@
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 export interface EthCallInterceptor {
   ethCall: (obs: Observable<any>) => Observable<any>;
