@@ -1,7 +1,6 @@
 import { ActionsObservable } from 'redux-observable';
 import { map, retry, catchError } from 'rxjs/operators';
-import { Observable } from 'rxjs/Observable';
-import { of } from 'rxjs/observable/of';
+import { Observable, of } from 'rxjs';
 
 import { EpicContext } from '../../context';
 import * as actions from './actions';
