@@ -1,4 +1,4 @@
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { EthProxyInterceptors } from './interceptors';
 import { ContractSchemaResolver } from './modules/schema';
 import { Provider, FilterObject, BlockchainEvent } from '@eth-proxy/rx-web3';
