@@ -1,10 +1,10 @@
 import { curry } from 'ramda';
 import * as Web3 from 'web3';
 import { createWeb3, bind } from '../utils';
-import { bindNodeCallback } from 'rxjs/observable/bindNodeCallback';
+import { bindNodeCallback } from 'rxjs';
 
 import { CurriedFunction2 } from 'ramda';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { BigNumber } from 'bignumber.js';
 
 export const getBalance = curry((provider: Web3.Provider, account: string) => {
