@@ -1,9 +1,8 @@
 import { curry, evolve } from 'ramda';
 import * as Web3 from 'web3';
 import { createWeb3, bind } from '../utils';
-import { bindNodeCallback } from 'rxjs/observable/bindNodeCallback';
 import { CurriedFunction2 } from 'ramda';
-import { Observable } from 'rxjs/Observable';
+import { bindNodeCallback, Observable } from 'rxjs';
 
 // It should be probably done by web3, same way like rest of receipt, but as for (0.20.3) its not done
 const decodeStatus = (status: string) =>
