@@ -1,9 +1,8 @@
 import { curry } from 'ramda';
 import * as Web3 from 'web3';
 import { createWeb3, bind } from '../utils';
-import { bindNodeCallback } from 'rxjs/observable/bindNodeCallback';
 import { CurriedFunction2 } from 'ramda';
-import { Observable } from 'rxjs/Observable';
+import { bindNodeCallback, Observable } from 'rxjs';
 
 export const getTransaction = curry((provider: Web3.Provider, hash: string) => {
   const web3 = createWeb3(provider);
